src/requirements.txt
tqdm==4.19.7
numpy==1.14.0
pyyaml==3.12.0

# Add the local PYPI server
--trusted-host dev-web.hutoma.com
--extra-index-url http://dev-web.hutoma.com:8088

# add dependencies
hu_logging==0.3.20
hu_utils==0.3.19
hu_training==0.3.20
